# Significant device enumeration speedup by using udev info instead of opening device files