This adds several early outs to the scanning before opening device files, if udev can returns information about devices, to avoid long synchronous device closes
* in `IsJoystick` if `SDL_UDEV_GetProductInfo` says device file is not joystick class
* in `IsSensor` if  `SDL_UDEV_GetProductInfo` says device file is not accelerometer class
* in `LINUX_ScanSteamVirtualGamepads` if `SDL_UDEV_GetProductInfo` says vendor and product isn't a Steam virtual gamepad

## Description

In #9092 it was revealed that closing device files can take anywhere from 0.5-0.01s. If the user has permissions to open a large number of device files, this can accumulate to significant startup time.

This resolves this in the case that the udev device data is available by using the udev devices data, which comes from _/run/udev_ or _/sys_ files with close times of 0.00002s (500-2,5000x faster), instead of opening devices when possible.

Here are profile and offcputime flamegraphs provided by @parkerlreed showing behaviour before this patch
* [offcputime](https://github.com/libsdl-org/SDL/assets/841440/32d544f0-02d7-4f1d-a14f-c20b81305e07)
* [profile](https://github.com/libsdl-org/SDL/assets/841440/18505936-1b43-43f1-b72b-bc4886477816)

and after it
* [offcputime](https://github.com/libsdl-org/SDL/assets/841440/72b138f9-62d8-4165-9881-aac8d34da41b)
* [profile](https://github.com/libsdl-org/SDL/assets/841440/c9f2f80c-816e-40c4-be84-b73980f9ba9b)

Note how total offcputime (time the program is blocked) went for over 2s to under 14ms (this varied a bit in subsequent test, but report are consistently under 300ms). Also note how the prior offcputime was dominated by `close` calls under `SDL_JoystickInit`, while the new one is now dominated by display related calls.

## Existing Issue(s)

See discussion at bottom of #9092. Note that 9092 was a combination of this slow device behaviour and commit 3b1e0e163ba3933daa9aa19f06a7bb3909e05c8a introducing O(n²) device walking for the full class of devices. The O(n²) device walking was resolved in that issue, but the general slowness was not.