@sezero 2024-04-25
```
/home/runner/work/SDL/SDL/src/joystick/linux/SDL_sysjoystick.c: In function ‘LINUX_ScanSteamVirtualGamepads’:
/home/runner/work/SDL/SDL/src/joystick/linux/SDL_sysjoystick.c:915:9: error: unused variable ‘class’ [-Werror=unused-variable]
  915 |     int class;
      |         ^~~~~
cc1: all warnings being treated as errors
```

I think the following needs applying to all branches

```diff
diff --git a/src/joystick/linux/SDL_sysjoystick.c b/src/joystick/linux/SDL_sysjoystick.c
index cdaee0b..bb7c31c 100644
--- a/src/joystick/linux/SDL_sysjoystick.c
+++ b/src/joystick/linux/SDL_sysjoystick.c
@@ -905,11 +905,13 @@ static void LINUX_ScanSteamVirtualGamepads(void)
     char path[PATH_MAX];
     struct input_id inpid;
     int num_virtual_gamepads = 0;
     int virtual_gamepad_slot;
     VirtualGamepadEntry *virtual_gamepads = NULL;
+#ifdef SDL_USE_LIBUDEV
     int class;
+#endif
 
     count = scandir("/dev/input", &entries, filter_entries, NULL);
     for (i = 0; i < count; ++i) {
         (void)SDL_snprintf(path, SDL_arraysize(path), "/dev/input/%s", entries[i]->d_name);
 
```