@twhitehead 2024-04-26
@slouken you will want to cherry pick the earlier commit  as well for SDL3 (there were two in this pull request).

It moves loading the udev library up a bit.

Without it, udev functions won't be available during the early manual scan, so will fall back to slow device opening.